+++ Pods/Headers/Private/Masonry/View+MASShorthandAdditions.h
@@ -1,0 +1,1 @@
+../../../../Masonry/View+MASShorthandAdditions.h