+++ Pods/Headers/Public/Masonry/MASViewAttribute.h
@@ -1,0 +1,1 @@
+../../../../Masonry/MASViewAttribute.h